% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr_diet_output_mean.R
\name{simmr_diet_output_mean}
\alias{simmr_diet_output_mean}
\title{Pull the average diets from simmr to run through Rpath}
\usage{
simmr_diet_output_mean(simmr_data, rpath_groups)
}
\arguments{
\item{simmr_data}{the output from running a simmr model. For more infomration on this, please see the simmr documentation.}

\item{rpath_groups}{a vector of all the groups included in the rpath model. This is crucial because it will assign a percentage of 0 to every species that is in the Rpath model, but not in the simmr model so the data is formatted in a way to be passed to Rpath.}
}
\value{
a list with the mean diet contribution for each species within the Rpath model.
}
\description{
This function pulls the means from the simmr diet output and formats them to be run through Rpath. This is intended to create a streamlined process to format diets from simmr to Rpath and run an initial model to check for balance.
}
\examples{
simmr_diet_output_mean(simmr_output, rpath_groups)
}
